Extract editor tab model from store
